requirements.txt
fastapi==0.110.0
python-jose==3.4.0
uvicorn==0.29.0
sqlalchemy==2.0.39
alembic==1.13.1
pydantic==2.10.6
bcrypt==4.1.2
pyjwt==2.8.0
passlib==1.7.4
python-dotenv==1.0.1
redis==5.0.4
psycopg2-binary==2.9.10
email-validator==2.2.0